Aaaand... remove strip-indent dependency
It didn't play nice with Browserify/bubleify transform. Its code didn't get converted to ES5 on the build. Could probably fix this by using Bublé directly over the Browserify bundle... but this is such a small dependency that i decided to just copy its code (for the time being :P)