Fix nasty parser bug caused by unintended global variable
The bug could be triggered by trying to parse this unbalanced-paren code:

  id = λx.x
  foo = id (id

Then fixing the issue by closing that parentheses or deleting it, and
then trying to parse it again.

The result was the parser complaining about an unexpected `=` in the
first line, because the lexer was retaining the `openParens` global
state from the previous parse() run.

Instead of tracking the open parentheses on a global variable, now the
code uses the `yy` Jison scope to store, making parse() runs independent
from each other, as they should be :)